+++ shift_lab/models/multitask_segformer/inference.py
@@ -10,8 +10,8 @@ from shift_dev.types import Keys
 from shift_dev.utils.backend import FileBackend
 from shift_dev.utils.load import im_decode
 
-from shift_lab.models.segformer.constants import SegformerTask
-from shift_lab.models.segformer.model import MultitaskSegformer
+from shift_lab.models.multitask_segformer.constants import SegformerTask
+from shift_lab.models.multitask_segformer.model import MultitaskSegformer
 
 
 class SHIFTInferenceCameraFrame:
